FileInfo: regularize the types of some fields. Uid, Gid become int. File size info becomes int64. Times become int64.
R=rsc, cw
CC=golang-dev
https://golang.org/cl/968042

+++ src/pkg/archive/tar/common.go
@@ -33,8 +33,8 @@ const (
 type Header struct {
 	Name     string
 	Mode     int64
-	Uid      int64
-	Gid      int64
+	Uid      int
+	Gid      int
 	Size     int64
 	Mtime    int64
 	Typeflag byte
